NI100X: Eliminate VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10855 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/NI100X/NI100X.cpp
@@ -542,7 +542,7 @@ long AnalogIO::GetListIndex()
    long listIdx(0);
    int ret = GetProperty(g_PropertyListIdx, listIdx);
    assert(ret == DEVICE_OK);
-   assert(listIdx < lists_.size());
+   assert(listIdx < static_cast<long>(lists_.size()));
    return listIdx;
 }
